Proofreading/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Proofreading"
version = "0.1.0"
dependencies = [
    "google-ai-generativelanguage==0.6.10",
    "google-generativeai==0.8.3",
    "google-adk==1.14.0",
    "python-dotenv==1.0.1",
    "click>=8.1.8",
    "mcp>=1.8.0",
    "httpx>=0.28.1",
    "markdown==3.7",
    "pathspec==0.12.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "chardet>=5.2.0",
    "pyyaml>=6.0.1",
]

requires-python = ">=3.13"

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "pylint>=3.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ["py313"]
include = "\\.pyi?$"
extend-exclude = '''
/(
    __pycache__
    | \.git
    | \.venv
    | build
    | dist
)/
'''

[tool.pylint.main]
max-line-length = 88
disable = [
    "C0103", # invalid-name
    "R0903", # too-few-public-methods
    "R0913", # too-many-arguments
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
